Separated into Model and View
